+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-	"editor.defaultFormatter": "biomejs.biome-nightly"
+	"editor.defaultFormatter": "biomejs.biome"
 }
