+++ src/auth/cookies.ts
@@ -1,4 +1,4 @@
 import { env } from '#server/env'
 
 export const USER_TOKEN =
-	env.NODE_ENV === 'production' ? '__Host-userToken' : 'userToken'
+	env.NODE_ENV === 'production' && !env.CI ? '__Host-userToken' : 'userToken'
